SWARM-1210: gradle related elements pulled out from the prod build
