+++ arquillian/adapter/pom.xml
@@ -100,12 +100,6 @@
     </dependency>
 
     <dependency>
-      <groupId>org.gradle</groupId>
-      <artifactId>gradle-tooling-api</artifactId>
-      <version>3.0</version>
-    </dependency>
-
-    <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <scope>compile</scope>
@@ -124,16 +118,5 @@
     </dependency>
 
   </dependencies>
-
-
-  <repositories>
-    <repository>
-      <id>gradle-tooling</id>
-      <url>http://repo.gradle.org/gradle/libs-releases-local</url>
-      <releases>
-        <enabled>true</enabled>
-      </releases>
-    </repository>
-  </repositories>
 
 </project>
