+++ arquillian/adapter/src/main/java/org/wildfly/swarm/arquillian/adapter/MavenDependencyDeclarationFactory.java
@@ -18,6 +18,8 @@ package org.wildfly.swarm.arquillian.adapter;
 import org.jboss.shrinkwrap.resolver.api.maven.MavenResolvedArtifact;
 import org.jboss.shrinkwrap.resolver.api.maven.PomEquippedResolveStage;
 import org.wildfly.swarm.arquillian.resolver.ShrinkwrapArtifactResolvingHelper;
+import org.wildfly.swarm.internal.FileSystemLayout;
+import org.wildfly.swarm.internal.MavenFileSystemLayout;
 import org.wildfly.swarm.tools.ArtifactSpec;
 import org.wildfly.swarm.tools.DeclaredDependencies;
 
@@ -25,7 +27,7 @@ import org.wildfly.swarm.tools.DeclaredDependencies;
  * @author Heiko Braun
  * @since 26/10/2016
  */
-class MavenDependencyDeclarationFactory extends DependencyDeclarationFactory {
+public class MavenDependencyDeclarationFactory implements DependencyDeclarationFactory {
 
 
     @Override
@@ -85,5 +87,10 @@ class MavenDependencyDeclarationFactory extends DependencyDeclarationFactory {
         }
 
         return declaredDependencies;
+    }
+
+    @Override
+    public boolean acceptsFsLayout(FileSystemLayout fsLayout) {
+        return fsLayout instanceof MavenFileSystemLayout;
     }
 }
