+++ arquillian/gradle-adapter/src/main/resources/META-INF/services/org.wildfly.swarm.arquillian.adapter.DependencyDeclarationFactory
@@ -1,0 +1,1 @@
+org.wildfly.swarm.arquillian.adapter.gradle.GradleDependencyDeclarationFactory
