+++ arquillian/pom.xml
@@ -30,4 +30,18 @@
     <module>test</module>
   </modules>
 
+  <profiles>
+    <profile>
+      <id>unsupported</id>
+      <activation>
+        <property>
+          <name>!swarm.product.build</name>
+        </property>
+      </activation>
+      <modules>
+        <module>gradle-adapter</module>
+      </modules>
+    </profile>
+  </profiles>
+
 </project>
