+++ gradle-plugin/src/main/java/org/wildfly/swarm/plugin/gradle/PackagePlugin.java
