+++ gradle-plugin/src/main/java/org/wildfly/swarm/plugin/gradle/PackageTask.java
