+++ gradle-plugin/src/main/resources/META-INF/gradle-plugins/wildfly-swarm.properties
